+++ packages/components-react/src/routes.tsx
@@ -396,6 +396,11 @@ export const routes: RouteType[] = [
     element: <fromExamples.MultiSelectExampleDynamicPage />,
   },
   {
+    name: 'Multi-Select Example Dynamic Optgroup',
+    path: '/multi-select-example-dynamic-optgroup',
+    element: <fromExamples.MultiSelectExampleDynamicOptgroupPage />,
+  },
+  {
     name: 'Multi-Select Example Form',
     path: '/multi-select-example-form',
     element: <fromExamples.MultiSelectExampleFormPage />,
@@ -476,6 +481,11 @@ export const routes: RouteType[] = [
     element: <fromExamples.SelectExampleDynamicPage />,
   },
   {
+    name: 'Select Example Dynamic Optgroup',
+    path: '/select-example-dynamic-optgroup',
+    element: <fromExamples.SelectExampleDynamicOptgroupPage />,
+  },
+  {
     name: 'Select Example Form',
     path: '/select-example-form',
     element: <fromExamples.SelectExampleFormPage />,
@@ -544,5 +554,5 @@ export const routes: RouteType[] = [
     name: 'Toast Example',
     path: '/toast-example',
     element: <fromExamples.ToastExamplePage />,
-  }
+  },
 ];