+++ packages/components/src/components/multi-select/multi-select/multi-select.spec.ts
@@ -69,6 +69,18 @@ describe('disconnectedCallback', () => {
   });
 });
 
+describe('optgroupUpdateHandler', () => {
+  it('should call stopPropagation() and updateOptions() with correct parameters', () => {
+    const component = initComponent();
+    const event = new Event('internalOptgroupUpdate', { bubbles: true });
+    const stopPropagationSpy = vi.spyOn(event, 'stopPropagation');
+
+    component.optgroupUpdateHandler(event);
+
+    expect(stopPropagationSpy).toHaveBeenCalled();
+  });
+});
+
 describe('formResetCallback', () => {
   it('should reset value to defaultValue', () => {
     const component = initComponent();