+++ packages/components/src/components/multi-select/multi-select/multi-select.tsx
@@ -170,6 +170,12 @@ export class MultiSelect {
     this.updateSelectedOption(e.target);
   }
 
+  @Listen('internalOptgroupUpdate')
+  public optgroupUpdateHandler(e: Event): void {
+    e.stopPropagation();
+    this.updateOptions();
+  }
+
   @Watch('value')
   public onValueChange(): void {
     this.setFormValue(this.value);