+++ packages/components/src/components/select/select/select.tsx
@@ -178,6 +178,12 @@ export class Select {
     this.updateSelectedOption(e.target);
   }
 
+  @Listen('internalOptgroupUpdate')
+  public optgroupUpdateHandler(e: Event): void {
+    e.stopPropagation();
+    this.updateOptions();
+  }
+
   @Watch('value')
   public onValueChange(): void {
     this.internals?.setFormValue(this.value);